ks-1-4 changed short to byte

+++ src/main/kotlin/lesson1/lesson1_task4.kt
@@ -2,7 +2,7 @@ package org.example.lesson1
 
 fun main(){
     val length: Long  = 40_868_600_000
-    val age: Short = 27
+    val age: Byte = 27
     val partOfDay: Float = 0.075F
     val seconds: Short = 6480
     val partOfYear: Double = 2.0547945205479453E-4
